third_party/blink/web_tests/svg/custom/viewbox-syntax-expected.txt

CONSOLE ERROR: Error: <svg> attribute viewBox: Unexpected end of attribute. Expected number, "0".
CONSOLE ERROR: Error: <svg> attribute viewBox: Unexpected end of attribute. Expected number, "0 0".
CONSOLE ERROR: Error: <svg> attribute viewBox: Unexpected end of attribute. Expected number, "0 0 30".
CONSOLE ERROR: Error: <svg> attribute viewBox: Trailing garbage, "0 0 30 40 50".
CONSOLE ERROR: Error: <svg> attribute viewBox: Trailing garbage, "0 0 30 40,".
CONSOLE ERROR: Error: <svg> attribute viewBox: Expected number, ",0 0 30 40".
CONSOLE ERROR: Error: <svg> attribute viewBox: A negative value is not valid. ("0 0 -30 40")
CONSOLE ERROR: Error: <svg> attribute viewBox: A negative value is not valid. ("0 0 30 -40")
CONSOLE ERROR: Error: <svg> attribute viewBox: Expected number, "0 0 30, , 40".
CONSOLE ERROR: Error: <svg> attribute viewBox: Expected number, "0 0 30% 40".
